// src/interfaces/CharacterInterface.ts
export interface CharacterInterface {
  birth_year: string;
  created: string;
  edited: string;
  eye_color: string;
  films: Array<string>;
  gender: string;
  hair_color: string;
  height: string;
  homeworld: string;
  mass: string;
  name: string;
  skin_color: string;
  species: Array<any>;
  starships: Array<string>;
  url: string;
  vehicles: Array<string>;
}